CSE: throw tar error on failure (#6200)

+++ vhdbuilder/packer/windows/configure-windows-vhd.ps1
@@ -503,6 +503,9 @@ function Install-ContainerD
     else
     {
         tar -xzf $containerdTmpDest -C $installDir
+        if ($LASTEXITCODE -ne 0) {
+            throw "Failed to extract the '$containerdTmpDest' archive."
+        }
         mv -Force $installDir\bin\* $installDir
         Remove-Item -Path $installDir\bin -Force -Recurse
     }
